add github action

+++ internal/regex_rate_limiter_test.go
@@ -110,7 +110,7 @@ regexes_with_rates:
 
 	nowNanos := float64(time.Now().UnixNano())
 	nowSeconds := nowNanos / 1e9
-	lineTime := fmt.Sprintf("%d", uint64(nowSeconds))
+	lineTime := fmt.Sprintf("%f", nowSeconds)
 	line := tail.Line{Text: lineTime + " 1.2.3.4 GET http://example.com/whatever " +
 		"AppleWebKit/537.36 (KHTML, like Gecko) Chrome/51.0.2704.103 Safari/537.36 -"}
 	fmt.Println("-- 1 --")
@@ -132,7 +132,7 @@ regexes_with_rates:
 	}
 
 	// 4 seconds after the first one
-	lineTime = fmt.Sprintf("%d", uint64(nowSeconds+4))
+	lineTime = fmt.Sprintf("%f", nowSeconds+4)
 	line = tail.Line{Text: lineTime + " 1.2.3.4 GET http://example.com/whatever " +
 		"AppleWebKit/537.36 (KHTML, like Gecko) Chrome/51.0.2704.103 Safari/537.36 -"}
 	fmt.Println("-- 2 --")
@@ -154,7 +154,7 @@ regexes_with_rates:
 	}
 
 	// a bit more than 5 seconds after the first one
-	lineTime = fmt.Sprintf("%d", uint64(nowSeconds+5.5))
+	lineTime = fmt.Sprintf("%f", nowSeconds+5.5)
 	line = tail.Line{Text: lineTime + " 1.2.3.4 GET http://example.com/whatever " +
 		"AppleWebKit/537.36 (KHTML, like Gecko) Chrome/51.0.2704.103 Safari/537.36 -"}
 	fmt.Println("-- 3 --")
@@ -176,7 +176,7 @@ regexes_with_rates:
 	}
 
 	// 1 second after the most recent one, but a POST instead of GET
-	lineTime = fmt.Sprintf("%d", uint64(nowSeconds+6.5))
+	lineTime = fmt.Sprintf("%f", nowSeconds+6.5)
 	line = tail.Line{Text: lineTime + " 1.2.3.4 POST http://example.com/whatever " +
 		"AppleWebKit/537.36 (KHTML, like Gecko) Chrome/51.0.2704.103 Safari/537.36 -"}
 	fmt.Println("-- 4 --")
@@ -205,7 +205,7 @@ regexes_with_rates:
 	}
 
 	// half a second after the most recent one, should exceed the rate limit
-	lineTime = fmt.Sprintf("%d", uint64(nowSeconds+7.0))
+	lineTime = fmt.Sprintf("%f", nowSeconds+7.0)
 	line = tail.Line{Text: lineTime + " 1.2.3.4 POST http://example.com/whatever " +
 		"AppleWebKit/537.36 (KHTML, like Gecko) Chrome/51.0.2704.103 Safari/537.36 -"}
 	fmt.Println("-- 5 --")
@@ -266,7 +266,7 @@ regexes_with_rates:
 
 	nowNanos := float64(time.Now().UnixNano())
 	nowSeconds := nowNanos / 1e9
-	lineTime := fmt.Sprintf("%d", uint64(nowSeconds))
+	lineTime := fmt.Sprintf("%f", nowSeconds)
 	line := tail.Line{Text: lineTime + " 1.2.3.4 GET http://skiphost.com/whatever " +
 		"AppleWebKit/537.36 (KHTML, like Gecko) Chrome/51.0.2704.103 Safari/537.36 -"}
 	fmt.Println("-- 1 --")
